+++ tzc_sales_customization_spt/models/stock_move.py
@@ -11,7 +11,7 @@ class stock_move(models.Model):
     _inherit = 'stock.move'
 
     package_id = fields.Many2one('kits.package.product',compute="_compute_package_id",store=True,compute_sudo=True)
-    qty_available = fields.Float('Total Qty',related='product_id.qty_available')
+    qty_available = fields.Integer('Total Qty',related='product_id.qty_available')
     pro_primary_image_url = fields.Char("Primary Image URL",related='product_id.primary_image_url')
     scan_extra_item = fields.Boolean(default=False)
     return_order_line_id = fields.Many2one('kits.return.ordered.items.line', string='Return Order Line')
